Make hierarchical-env example recognized by cargo
