+++ Cargo.toml
@@ -39,6 +39,7 @@ json5_rs = { version = "0.3", optional = true, package = "json5" }
 indexmap = { version = "1.7.0", features = ["serde-1"], optional = true}
 
 [dev-dependencies]
+serde = "1.0.8"
 serde_derive = "1.0.8"
 float-cmp = "0.8"
 chrono = { version = "0.4", features = ["serde"] }
