--- examples/hierarchical-env/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "hierarchical-env"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-config = { path = "../../" }
-serde_derive = "^1.0.8"
-serde = "^1.0.8"
